t19x: galen: Add regulator-fixed compatible
The "regulator-fixed-sync" is a downstream implementation that is
not supported by kstable tree where we need to keep the code line
as closer to upstream as possible.

So this patch adds a "regulator-fixed", as a secondary compatible
string, to make it work for kstable tree.

TMM-1934

Change-Id: Ie2cdd9512456b440376dbf549aeb82516310ae70
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/stardust-dts/+/2345323
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Kartik . <kkartik@nvidia.com>
Reviewed-by: Rohit Khanna <rokhanna@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>