video: tegra: host: Change v2 DMA fence refcounting
Don't keep a reference to the DMA fence until an interrupt is
triggered, as we drop the intr ref anyway when releasing the
fence. This also allows dropping userspace-created fences even
if they weren't signaled, which is good. It also fixes an
intermittent issue when running looped timeout tests.

Change-Id: I3e0be7262448c6e405a7dabdb3236f09c4c3954a
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2361646
GVS: Gerrit_Virtual_Submit
Reviewed-by: Konsta Holtta <kholtta@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>