video: tegra: update bool check in SELECT_T_PHY
In Kernel-5.9, compiler complains if equations are used as
boolean checks.
E.g.: In some cases, the value passed to 'platform_t_phy_ps'
parameter of SELECT_T_PHY is an equation like 'xyz * 1000'.
This fails to build with "error: ‘*’ in boolean context".

To overcome this, updated the check to compare with 0 instead.

bug 200601926

Change-Id: I4e6d991e3654c0f38086316552fa04cdb91eb0b8
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2365679
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>