gpu: nvgpu: fix posix cmpxchg()
Commit bd1ae5c9e1e5 ("gpu: nvgpu: fix MISRA 17.7 violations in mm") did
a seemingly harmless looking change in the cmpxchg() wrapper macro to
convert from atomic_compare_exchange_strong() to nvgpu_atomic_cmpxchg().
The latter is ultimately a wrapper for the former but the semantics are
different: the former takes old as a pointer and updates it for the read
value, while the latter takes it as a value and returns the read value.
The commit caused cmpxchg() to always return the old value, so a failing
compare has never been detected in a year and half.

This cmpxchg() is used only in the lockless allocator which is used only
in the fence code in deterministic kernel submits which hasn't been part
of safe code, so the broken code has been basically not used. (The
typecast from an integer pointer to an atomic pointer is a separate
concern.)

Jira NVGPU-5493

Change-Id: I932a69c6c185783c0e514e848e0ee6057ce74888
Signed-off-by: Konsta Hölttä <kholtta@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvgpu/+/2368118
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Alex Waterman <alexw@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>