video: tegra: host: nvdla: Fix MISRA 10.3 violation
The constants defined in dla_*_fw_version.h have UL suffix which
translates to 64 bit unsigned integer on 64 bit architectures and 32
bit unsigned integers on 32 bit architectures such as falcon.
To be MISRA rule 10.3 compliant on both, use explicit typecast to
uint32_t which should work across both.

This change does not fix any violations in here but reflects an
update to this file that helps fix 10.3 violation in QNX KMD.

Change-Id: Ia47c2f97ee8c18bfeea071861d5dbab15293551a
Signed-off-by: Anup Mahindre <amahindre@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-t23x/+/2369151
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>