t23x: icc: replace mc client id's with abstract id
MC client ID's are not contiguous. This creates a problem when these
MC client ID's are used as ICC ID's. of_icc_get() which is the device
tree method of getting icc_path handle requires the ID's to be less
than number of ICC nodes. To fix this problem replace the MC client
ID's with abstract ICC ID's which are contiguous and starting at
index 0. All the other ICC providers also make use of abstarct ID's
and hence this change will comply with that. Also the primary
motivation for using MC client ID's was to help with the
programming of MC registers. As the programming of these MC registers
has been moved out to BPMP-FW, this requirement does not apply
anymore.

Replace master/slave phrase usage in ICC with primary/secondary.

JIRA: TPM-1265

Change-Id: I39cba6337fe0e53fb4bb95f4f48d29906d6d44a2
Signed-off-by: Sanjay Chandrashekara <sanjayc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-t23x/+/2374870
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Sitaraman <ksitaraman@nvidia.com>
Reviewed-by: Vinod Gopalakrishnakurup <vinodg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit