TEMP: mailbox: tegra: hsp: skip hsp db check
On simulation platforms the BPMP hasn't booted up yet by the time we
probe the BPMP driver and the BPMP hasn't had a chance to mark the
doorbell as ringable by the CCPLEX. This corresponding check in the
BPMP driver will therefore fail. Work around this by disabling the
check on simulation platforms.

Revert this patch when similar changes are available from upstream
kernel.

Bug 200548503
Bug 200640055

Change-Id: I4e2effd901d048be211b517323ef29bc0ee8741d
Signed-off-by: Rajesh Gumasta <rgumasta@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2383515
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit