firmware: tegra: Allow for DRAM MRQ carveouts
Today the carveout used to communicate between BPMP and CPU-NS lives in SYSRAM.
This driver accesses it using the SYSRAM driver and the SRAM framework.
However for Tegra234, this carveout will move to DRAM and the location will be
signaled using the reserved memory node in DT.
This patch adapts the driver to deal with the upcoming changes.

Change-Id: I21267fc43a2ebc5ea50d61d3ca5f15a327665bc8
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2385373
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit