video: tegra: host: reset max for client managed on free
When freeing a syncpoint we currently set the min value equal to the max
value. This works as intended for host-managed syncpoints, but
client-managed might have their max values untouched which could rewind
the syncpoint back to what it was when allocated. So do the inverse and
reset max to min.

This logic relies on that the client must have stopped using the
syncpoint in hardware so that the min value has settled and won't change
anymore, so the current min value is a reasonable new max.

Jira NVGPU-5506

Change-Id: I08b3c760c1212aa478e028117bf14b4eab9becff
Signed-off-by: Konsta Hölttä <kholtta@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2400853
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>