gpu: nvgpu: Fix building NVGPU when NVHost is a module
If NVHost is built as a module then NVGPU is built without support
for NVHost because the Makefile for NVGPU is only enabling NVHost
support when NVHost is built into the kernel. Fix this by enabling
support in NVGPU for NVHost when NVHost is built into the kernel or
as a module.

Bug 200617256

Change-Id: I7226daa8b4815147a465d5b3fd3ac3e3cdae8d36
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvgpu/+/2401332
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Debarshi Dutta <ddutta@nvidia.com>
Reviewed-by: Sagar Kamble <skamble@nvidia.com>
Reviewed-by: Alex Waterman <alexw@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit