video: tegra: host: Add sync file extract IOCTL
The old Android sync FD IOCTLs exposed a way to copy driver-specific
data to userspace through the sync FD IOCTLs. This was used to convert
sync FDs to id/threshold pairs for the userspace. With sync files,
this is no longer there, so add an NvHost IOCTL for the same purpose.

Change-Id: I0d4f7f8e5ac5aec611886a9154b362ba03f89259
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2402033
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Konsta Holtta <kholtta@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>