ALSA: hda/tegra:  WAR for GCAP ISS value on T23x
The GCAP register on T23x implies no Input Streams(ISS) supported,
but the HW output stream descriptor programming should start with
offset 0x20*4 from base stream descriptor address. This will be a
problem while calculating the offset for output stream descriptor
which will be considering input stream also. So here output stream
starts with offset 0 which is wrong as HW register for output stream
offset starts with 4. So hardcode the input stream numbers to 4 to
avoid the issue in offset calculation.

Bug 200613795

Change-Id: Iddb39576544c238d4d8fff8d85d6ae38369d80d8
Signed-off-by: Mohan Kumar <mkumard@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2402877
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>